#52 #28 test for Write_Ini_for_CFG in temp dir. And some refactoring for Repo_File_Paths

+++ project/test_svn_wrapper.py
@@ -17,9 +17,10 @@ class SVN_Wrapper_Abstract_Test(TestCase):
         repo_name = 'repo_name'
         t = Repo_File_Paths( repo_root, repo_name )
 
-        self.assertEqual( t.svnserve_conf_full_name( ), repo_root + repo_name + '\\conf\\svnserve.conf' )
+        self.assertEqual( t.conf_folder( ), repo_root + '\\' + repo_name + '\\conf\\' )
+        self.assertEqual( t.svnserve_conf_full_name( ), repo_root + '\\' + repo_name + '\\conf\\svnserve.conf' )
         self.assertEqual( t.pass_full_name( ),          repo_root + '\\passwd' ) # one passwd in root
-        self.assertEqual( t.auth_full_name( ),          repo_root + repo_name + '\\conf\\authz' )
+        self.assertEqual( t.auth_full_name( ),          repo_root + '\\' + repo_name + '\\conf\\authz' )
 
     def test_Repo_PW_Encoing(self):
         test_pw = 'some test pass'
@@ -34,17 +35,36 @@ class SVN_Wrapper_Temp_Dir_Test(TestCase):
 
     def setUp(self):
         # Create a temporary directory
-        self.test_temp_dir = tempfile.mkdtemp()
+        if not self.test_temp_dir:
+            self.test_temp_dir = tempfile.mkdtemp()
 
     def tearDown(self):
         # Remove the directory after the test
-        shutil.rmtree(self.test_temp_dir)
+        if self.test_temp_dir:
+            shutil.rmtree(self.test_temp_dir)
+            self.test_temp_dir = None
+
+    def test_Repo_Users_In_Temp_Dir(self):
+        repo_root = self.test_temp_dir
+        repo_name = 'repo_name'
+        t = Repo_File_Paths( repo_root, repo_name )
+        fn = t.pass_full_name()
+
+        Add_User_To_Main_PassFile( fn, { 'user_name1' : 'pass', 'user_name2' : 'pass2' } )
+        f = open( fn )
+        self.assertEqual(f.read(), '[users]\nuser_name1 = pass\nuser_name2 = pass2\n\n')
+        f.close()
 
     def test_Repo_Conf_In_Temp_Dir(self):
         repo_root = self.test_temp_dir
         repo_name = 'repo_name'
         t = Repo_File_Paths( repo_root, repo_name )
+        fn = t.svnserve_conf_full_name()
 
-        Add_User_To_Main_PassFile( t.pass_full_name(), { 'user_name1' : 'pass', 'user_name2' : 'pass2' } )
-        f = open( t.pass_full_name() )
-        self.assertEqual(f.read(), '[users]\nuser_name1 = pass\nuser_name2 = pass2\n\n')
+        import os
+        os.makedirs( t.conf_folder() ) # force path to cfg
+
+        Write_Ini_for_CFG( fn, 'some_section', { 'option' : 'value' } )
+        f = open( fn )
+        self.assertEqual(f.read(), '[some_section]\noption = value\n\n')
+        f.close()
